refactor: added folders for interfaces in application layer
