+++ src/Application/Interfaces/Services/General/IWebViewTools.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services;
+﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services.General;
 
 using System.Threading.Tasks;
