+++ src/Application/Interfaces/Services/Geo/ILocationService.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services;
+﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services.Geo;
 
 using System.Collections.Generic;
 using System.Threading;
