+++ src/Application/Interfaces/Services/Initiatives/IJoinRequestService.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services;
+﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services.Initiatives;
 
 using System.Threading;
 using System.Threading.Tasks;
