+++ src/Application/Interfaces/Services/Statistics/IGeneralStatisticsService.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services;
+﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services.Statistics;
 
 using System.Threading;
 using System.Threading.Tasks;
