+++ src/Application/Interfaces/Services/TerritoryStory/ITerritoryStoryImageService.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services;
+﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services.TerritoryStory;
 
 using System.Threading;
 using System.Threading.Tasks;
