+++ src/WebApi/Controllers/Rest/Geo/LocationController.cs
@@ -3,7 +3,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-using IAVH.BioTablero.CM.Application.Interfaces.Services;
+using IAVH.BioTablero.CM.Application.Interfaces.Services.Geo;
 using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples;
 using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.Location;
 using IAVH.BioTablero.CM.WebApi.Interfaces;
